[netstack3] Cleanup TCP established_send test
- Use the constant TEST_BYTES instead of the literal as it is shared
  among other tests.
- Make WindowSize constructor accept usize to remove unnecessary *_U32
  constants.

Test: No behavior change
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663688
Commit-Queue: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>

+++ src/connectivity/network/netstack3/core/src/transport/tcp/segment.rs
@@ -416,7 +416,7 @@ mod test {
         control: Option<Control>,
         data_len: u32,
         rcv_nxt: u32,
-        rcv_wnd: u32,
+        rcv_wnd: usize,
     }
     #[test_case(OverlapTestArgs{
         seg_seq: 1,
